fix, add test
