+++ temporian/implementation/numpy_cc/operators/window.cc
@@ -10,16 +10,19 @@
 #include <type_traits>
 #include <vector>
 
-using namespace std; // TODO: REMOVE + REMOVE ALL APPEARANCES OF cout <<
-
 namespace {
 namespace py = pybind11;
 
 typedef py::array_t<double> ArrayD;
 typedef py::array_t<float> ArrayF;
 
-// Apply TAccumulator over the data sequentially and get aggregated results
-// Use evset's timestamps and a fixed window_length
+// NOTE: accumulate() is overloaded for the 4 possible combinations of:
+// - with or without external sampling
+// - with constant or variable window length
+
+// TODO: refactor to avoid code duplication where possible.
+
+// No external sampling, constant window length
 template <typename INPUT, typename OUTPUT, typename TAccumulator>
 py::array_t<OUTPUT> accumulate(const ArrayD &evset_timestamps,
                                const py::array_t<INPUT> &evset_values,
@@ -77,8 +80,7 @@ py::array_t<OUTPUT> accumulate(const ArrayD &evset_timestamps,
   return output;
 }
 
-// Apply TAccumulator over the data sequentially and get aggregated results
-// Use external sampling_timestamps and a fixed window_length
+// External sampling, constant window length
 template <typename INPUT, typename OUTPUT, typename TAccumulator>
 py::array_t<OUTPUT> accumulate(const ArrayD &evset_timestamps,
                                const py::array_t<INPUT> &evset_values,
@@ -123,15 +125,11 @@ py::array_t<OUTPUT> accumulate(const ArrayD &evset_timestamps,
   return output;
 }
 
-// Apply TAccumulator over the data sequentially and get aggregated results
-// Use evset's timestamps and a variable window_length
+// No external sampling, variable window length
 template <typename INPUT, typename OUTPUT, typename TAccumulator>
 py::array_t<OUTPUT> accumulate(const ArrayD &evset_timestamps,
                                const py::array_t<INPUT> &evset_values,
                                const ArrayD &window_length) {
-
-  cout << "variable window length\n";
-
   // Input size
   const size_t n_event = evset_timestamps.shape(0);
 
@@ -142,18 +140,6 @@ py::array_t<OUTPUT> accumulate(const ArrayD &evset_timestamps,
   auto v_timestamps = evset_timestamps.unchecked<1>();
   auto v_values = evset_values.template unchecked<1>();
   auto v_window_length = window_length.unchecked<1>();
-
-  cout << "window length shape: " + to_string(v_window_length.shape(0)) + "\n";
-  cout << "window length first element: " + to_string(v_window_length[0]) +
-              "\n";
-  cout << "window length last element: " +
-              to_string(v_window_length[v_window_length.shape(0) - 1]) + "\n";
-
-  // same for v_values
-  cout << "values shape: " + to_string(v_values.shape(0)) + "\n";
-  cout << "values first element: " + to_string(v_values[0]) + "\n";
-  cout << "values last element: " + to_string(v_values[v_values.shape(0) - 1]) +
-              "\n";
 
   TAccumulator accumulator;
 
@@ -173,9 +159,6 @@ py::array_t<OUTPUT> accumulate(const ArrayD &evset_timestamps,
     curr_ts = v_timestamps[end_idx];
     curr_window_length = v_window_length[end_idx];
 
-    cout << "curr_ts: " + to_string(curr_ts) + "\n";
-    cout << "curr_window_length: " + to_string(curr_window_length) + "\n";
-
     // Add all values with same timestamp as the current one.
     accumulator.Add(v_values[end_idx]);
     size_t first_diff_ts_idx = end_idx + 1;
@@ -185,33 +168,34 @@ py::array_t<OUTPUT> accumulate(const ArrayD &evset_timestamps,
       first_diff_ts_idx++;
     }
 
-    bool begin_moved_forward =
-        end_idx > 0 &&
-        curr_ts - v_timestamps[end_idx - 1] -
-                (curr_window_length - v_window_length[end_idx - 1]) >
-            0;
-    cout << "begin_moved_forward: " + to_string(begin_moved_forward) + "\n";
+    if (end_idx > 0) {
+      bool begin_moved_forward =
+          curr_ts - v_timestamps[end_idx - 1] -
+              (curr_window_length - v_window_length[end_idx - 1]) >
+          0;
 
-    // Move begin_idx forward or backwards depending on begin_diff.
-    if (begin_moved_forward) {
-      // Window's beginning moved forward
-      while (begin_idx < n_event &&
-             v_timestamps[end_idx] - v_timestamps[begin_idx] >=
-                 curr_window_length) {
-        // cout << "Removing value from accumulator: " +
-        //             to_string(v_values[begin_idx]) + "\n";
-        accumulator.Remove(v_values[begin_idx]);
-        begin_idx++;
-      }
-    } else {
-      // Window's beginning moved backwards.
-      // Note < instead of <= to respect (] window boundaries.
-      while (begin_idx >= 0 && v_timestamps[end_idx] - v_timestamps[begin_idx] <
-                                   curr_window_length) {
-        // cout << "Adding value to accumulator: " +
-        //             to_string(v_values[begin_idx]) + "\n";
-        accumulator.Add(v_values[begin_idx]);
+      // Move begin_idx forward or backwards depending on begin_diff.
+      if (begin_moved_forward) {
+        // Window's beginning moved forward
+        while (begin_idx < n_event &&
+               v_timestamps[end_idx] - v_timestamps[begin_idx] >=
+                   curr_window_length) {
+          accumulator.Remove(v_values[begin_idx]);
+          begin_idx++;
+        }
+      } else {
+        // Window's beginning moved backwards.
+        // Note < instead of <= to respect (] window boundaries.
+        // Value at current begin_idx is already in the accumulator so we need
+        // to move begin_idx one step left, and then back one step right.
         begin_idx--;
+        while (begin_idx >= 0 &&
+               v_timestamps[end_idx] - v_timestamps[begin_idx] <
+                   curr_window_length) {
+          accumulator.Add(v_values[begin_idx]);
+          begin_idx--;
+        }
+        begin_idx++;
       }
     }
 
@@ -223,16 +207,12 @@ py::array_t<OUTPUT> accumulate(const ArrayD &evset_timestamps,
 
     // Move pointer to the index of the last value with the same timestamp.
     end_idx = first_diff_ts_idx;
-
-    cout << "\n\n";
   }
 
   return output;
 }
 
-// TODO
-// Apply TAccumulator over the data sequentially and get aggregated results
-// Use evset's timestamps and a variable window_length
+// External sampling, variable window length
 template <typename INPUT, typename OUTPUT, typename TAccumulator>
 py::array_t<OUTPUT> accumulate(const ArrayD &evset_timestamps,
                                const py::array_t<INPUT> &evset_values,
@@ -243,6 +223,8 @@ py::array_t<OUTPUT> accumulate(const ArrayD &evset_timestamps,
 
   // Allocate output array
   auto output = py::array_t<OUTPUT>(n_event);
+
+  // TODO (ian): implement
 
   return output;
 }
